+++ website/app/docs/Concurrency/_meta.js
@@ -1,4 +1,5 @@
 export default {
+  abortable: 'abortable',
   filterAsync: 'filterAsync',
   flatMapAsync: 'flatMapAsync',
   eachAsync: 'eachAsync',